change activity bar directory
